feat: adding code for release v0.15

+++ python/mod_cvcuda/OpSIFT.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2023-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 2023-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -54,7 +54,7 @@ public:
 
         long long int payloadSize() const
         {
-            return m_maxShape.x * m_maxShape.y * m_maxShape.z * m_maxOctaveLayers;
+            return static_cast<long long int>(m_maxShape.x) * m_maxShape.y * m_maxShape.z * m_maxOctaveLayers;
         }
 
     private:
